refactor: stop using start methods on other commands
Achieve this by moving the state saving and loading responsibility to the LocalState itself
