+++ linkup-cli/src/start.rs
@@ -1,5 +1,5 @@
 use std::{
-    fs::{self, File},
+    fs,
     path::{Path, PathBuf},
 };
 
@@ -10,12 +10,12 @@ use crate::{
     linkup_file_path,
     local_config::{config_to_state, LocalState, YamlLocalConfig},
     status::{server_status, ServerStatus},
-    CliError, LINKUP_STATE_FILE,
+    CliError,
 };
 use crate::{services, LINKUP_LOCALDNS_INSTALL};
 
 pub fn start(config_arg: &Option<String>) -> Result<(), CliError> {
-    let previous_state = get_state();
+    let previous_state = LocalState::load();
     let config_path = config_path(config_arg)?;
     let input_config = get_config(&config_path)?;
 
@@ -30,7 +30,7 @@ pub fn start(config_arg: &Option<String>) -> Result<(), CliError> {
         state.linkup.tunnel = ps.linkup.tunnel;
     }
 
-    save_state(state.clone())?;
+    state.save()?;
 
     // Set env vars to linkup
     for service in &state.services {
@@ -47,42 +47,6 @@ pub fn start(config_arg: &Option<String>) -> Result<(), CliError> {
     boot_background_services()?;
 
     check_local_not_started()?;
-
-    Ok(())
-}
-
-pub fn get_state() -> Result<LocalState, CliError> {
-    if let Err(e) = File::open(linkup_file_path(LINKUP_STATE_FILE)) {
-        return Err(CliError::NoState(e.to_string()));
-    }
-
-    let content = match fs::read_to_string(linkup_file_path(LINKUP_STATE_FILE)) {
-        Ok(content) => content,
-        Err(e) => return Err(CliError::NoState(e.to_string())),
-    };
-
-    match serde_yaml::from_str(&content) {
-        Ok(config) => Ok(config),
-        Err(e) => Err(CliError::NoState(e.to_string())),
-    }
-}
-
-pub fn save_state(state: LocalState) -> Result<(), CliError> {
-    let yaml_string = match serde_yaml::to_string(&state) {
-        Ok(yaml) => yaml,
-        Err(_) => {
-            return Err(CliError::SaveState(
-                "Failed to serialize the state into YAML".to_string(),
-            ))
-        }
-    };
-
-    if fs::write(linkup_file_path(LINKUP_STATE_FILE), yaml_string).is_err() {
-        return Err(CliError::SaveState(format!(
-            "Failed to write the state file at {}",
-            linkup_file_path(LINKUP_STATE_FILE).display()
-        )));
-    }
 
     Ok(())
 }
@@ -130,7 +94,7 @@ fn set_service_env(directory: String, config_path: String) -> Result<(), CliErro
 }
 
 fn check_local_not_started() -> Result<(), CliError> {
-    let state = get_state()?;
+    let state = LocalState::load()?;
     for service in state.services {
         if service.local == service.remote {
             continue;
